test: Split test_ai_user_id.py by concern; document the seam-first test policy
Split the 8,255-line file into two ResponseRunner lifecycle modules plus
a shared tests/ai_user_id_helpers.py, keeping the ai.py-seam class where
it was:

- test_response_runner_session_lifecycle.py  process_and_respond,
  generate_response_locked, session-started hooks, interrupted-history
  persistence (38 module tests)
- test_response_runner_team_streaming.py     team-helper streaming
  delivery failures, cancellation persistence, prompt merging (23)
- test_ai_user_id.py (slimmed)               10 metadata/prompt unit
  tests + TestUserIdPassthrough verbatim (the ai.py seam)

tests/test_locked_turn_delivery.py now imports its helpers from
tests/ai_user_id_helpers.py instead of reaching into a test module.

Also documents the ongoing policy in CLAUDE.md (Step 4): new behavior
tests target the collaborator seam that owns the behavior via the
conftest seam installers or *Deps dataclasses, never bot attribute
mocks; >3 patches per test is a wrong-seam smell; end-to-end tests live
in the slim integration files.

Tests moved verbatim (AST-extracted blocks). Collected count pinned:
169 before (test_ai_user_id + test_locked_turn_delivery), 169 after,
sorted names byte-identical; every file passes in isolation. Full suite
9112 passed / 61 skipped (9173 collected, unchanged).
