+++ tests/test_response_runner_team_streaming.py
@@ -1,0 +1,1599 @@
+"""ResponseRunner team-helper streaming lifecycle: delivery failures, cancellation persistence, and prompt merging."""
+
+from __future__ import annotations
+
+import asyncio
+from contextlib import suppress
+from dataclasses import replace
+from types import SimpleNamespace
+from typing import TYPE_CHECKING, cast
+from unittest.mock import AsyncMock, MagicMock, patch
+
+import pytest
+from agno.models.message import Message
+from agno.run.base import RunStatus
+from agno.run.team import TeamRunOutput
+from agno.session.team import TeamSession
+
+from mindroom.bot import AgentBot
+from mindroom.config.agent import AgentConfig, AgentPrivateConfig
+from mindroom.config.main import Config
+from mindroom.config.models import ModelConfig
+from mindroom.constants import (
+    ROUTER_AGENT_NAME,
+)
+from mindroom.final_delivery import FinalDeliveryOutcome, StreamTransportOutcome
+from mindroom.history.turn_recorder import TurnRecorder
+from mindroom.hooks import (
+    EVENT_SESSION_STARTED,
+    HookRegistry,
+    SessionHookContext,
+    hook,
+)
+from mindroom.message_target import MessageTarget
+from mindroom.response_runner import (
+    ResponseRunner,
+)
+from mindroom.streaming import StreamingDeliveryError
+from mindroom.tool_system.events import ToolTraceEntry
+from tests.identity_helpers import fixture_entity_matrix_id
+
+if TYPE_CHECKING:
+    from collections.abc import AsyncIterator, Awaitable, Callable
+    from pathlib import Path
+
+
+from tests.ai_user_id_helpers import (
+    _build_response_runner,
+    _config,
+    _config_with_team,
+    _config_with_team_matrix_message,
+    _handled_response_event_id,
+    _knowledge_access_support,
+    _make_bot,
+    _open_team_scope_context,
+    _plugin,
+    _response_request,
+    _runtime_paths,
+    _SessionStorage,
+    _set_gateway_method,
+    _stream_outcome,
+    _team_orchestrator,
+    bind_runtime_paths,
+)
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_preserves_raw_prompt_when_model_prompt_supplies_tail(
+    tmp_path: Path,
+) -> None:
+    """Team responses should keep the raw user text when model_prompt only adds transient tails."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch(
+            "mindroom.response_runner.team_response",
+            new=AsyncMock(return_value="Team answer"),
+        ) as mock_team_response,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        await coordinator.generate_team_response_helper(
+            replace(
+                _response_request(prompt="Describe this image", user_id="@alice:localhost", thread_id="$thread-root"),
+                model_prompt="Available attachment IDs: att_1. Use tool calls to inspect or process them.",
+            ),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    message = mock_team_response.await_args.kwargs["message"]
+    assert "Describe this image" in message
+    assert "Available attachment IDs: att_1" in message
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_appends_matrix_tool_prompt_context(tmp_path: Path) -> None:
+    """Team Matrix tool metadata appended during runtime preparation should reach the model."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team_matrix_message(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    model_messages: list[str] = []
+
+    async def fake_team_response(*_args: object, **kwargs: object) -> str:
+        model_messages.append(cast("str", kwargs["message"]))
+        return "Team answer"
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_runner.team_response", new=AsyncMock(side_effect=fake_team_response)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert model_messages
+    assert "[Matrix metadata for tool calls]" in model_messages[0]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_allows_explicit_private_ad_hoc_member(tmp_path: Path) -> None:
+    """ResponseRunner preflight should not reject direct private members before team_response."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(
+        Config(
+            agents={
+                "private_worker": AgentConfig(
+                    display_name="PrivateWorker",
+                    private=AgentPrivateConfig(per="user", root="private_worker_data"),
+                ),
+                "calculator": AgentConfig(display_name="Calculator"),
+            },
+            models={"default": ModelConfig(provider="openai", id="test-model")},
+        ),
+        runtime_paths,
+    )
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="calculator")
+    seen_agent_names: list[list[str]] = []
+
+    async def fake_team_response(*_args: object, **kwargs: object) -> str:
+        seen_agent_names.append(cast("list[str]", kwargs["agent_names"]))
+        return "Team answer"
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_runner.team_response", new=AsyncMock(side_effect=fake_team_response)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[
+                fixture_entity_matrix_id("private_worker", "localhost", runtime_paths),
+                fixture_entity_matrix_id("calculator", "localhost", runtime_paths),
+            ],
+            team_mode="coordinate",
+        )
+
+    assert seen_agent_names == [["private_worker", "calculator"]]
+    bot._conversation_state_writer.team_history_scope.assert_called_once_with(
+        [
+            fixture_entity_matrix_id("private_worker", "localhost", runtime_paths),
+            fixture_entity_matrix_id("calculator", "localhost", runtime_paths),
+        ],
+        requester_user_id="@alice:localhost",
+    )
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_passes_resolved_correlation_id_to_team_response(tmp_path: Path) -> None:
+    """Team execution should share the lifecycle/tool-runtime correlation id."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    seen_kwargs: dict[str, object] = {}
+
+    async def fake_team_response(*_args: object, **kwargs: object) -> str:
+        seen_kwargs.update(kwargs)
+        return "Team answer"
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_runner.team_response", new=AsyncMock(side_effect=fake_team_response)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$original"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        await coordinator.generate_team_response_helper(
+            _response_request(
+                prompt="Regenerate team edit",
+                user_id="@alice:localhost",
+                thread_id="$thread-root",
+                reply_to_event_id="$original",
+                correlation_id="$edit",
+            ),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    ctx = seen_kwargs["ctx"]
+    assert ctx.reply_to_event_id == "$original"
+    assert ctx.correlation_id == "$edit"
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_preserves_model_prompt_in_persisted_session(
+    tmp_path: Path,
+) -> None:
+    """Team persisted model prompts should stay intact for later provider cache reuse."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    async def fake_team_response(*_args: object, **kwargs: object) -> str:
+        model_message = cast("str", kwargs["message"])
+        run_id = cast("str | None", kwargs.get("run_id"))
+        storage.session = TeamSession(
+            session_id="!test:localhost:$thread-root",
+            team_id="ultimate",
+            created_at=1,
+            updated_at=1,
+            runs=[
+                TeamRunOutput(
+                    run_id=run_id,
+                    content="Team answer",
+                    messages=[
+                        Message(role="user", content="Earlier context"),
+                        Message(role="assistant", content="Earlier answer"),
+                        Message(role="user", content=model_message),
+                        Message(role="assistant", content="Team answer"),
+                    ],
+                ),
+            ],
+        )
+        return "Team answer"
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_runner.team_response", new=AsyncMock(side_effect=fake_team_response)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        await coordinator.generate_team_response_helper(
+            replace(
+                _response_request(prompt="Describe this image", user_id="@alice:localhost", thread_id="$thread-root"),
+                model_prompt="Available attachment IDs: att_1. Use tool calls to inspect or process them.",
+            ),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert persisted_run.messages is not None
+    assert persisted_run.messages[0].content == "Earlier context"
+    assert "Describe this image" in cast("str", persisted_run.messages[2].content)
+    assert "Available attachment IDs: att_1" in cast("str", persisted_run.messages[2].content)
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_preserves_retry_model_prompt(tmp_path: Path) -> None:
+    """Team retry runs should keep the model-facing prompt that Agno persisted."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+    seen_run_ids: list[str | None] = []
+
+    async def fake_team_response(*_args: object, **kwargs: object) -> str:
+        model_message = cast("str", kwargs["message"])
+        run_id = cast("str | None", kwargs["ctx"].run_id)
+        seen_run_ids.append(run_id)
+        run_id_callback = cast("Callable[[str], None]", kwargs["run_id_callback"])
+        if run_id is not None:
+            run_id_callback(run_id)
+        storage.session = TeamSession(
+            session_id="!test:localhost:$thread-root",
+            team_id="ultimate",
+            created_at=1,
+            updated_at=1,
+            runs=[
+                TeamRunOutput(
+                    run_id=run_id,
+                    content="Team answer",
+                    messages=[
+                        Message(role="user", content=model_message),
+                        Message(role="assistant", content="Team answer"),
+                    ],
+                ),
+            ],
+        )
+        return "Team answer"
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_runner.team_response", new=AsyncMock(side_effect=fake_team_response)),
+        patch(
+            "mindroom.teams.open_bound_scope_session_context",
+            side_effect=lambda **_kwargs: _open_team_scope_context(storage),
+        ),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        coordinator.deps.delivery_gateway.send_text.return_value = "$msg"
+
+        await coordinator.generate_team_response_helper(
+            replace(
+                _response_request(prompt="Describe this image", user_id="@alice:localhost", thread_id="$thread-root"),
+                model_prompt="Available attachment IDs: att_1. Use tool calls to inspect or process them.",
+            ),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert seen_run_ids == [persisted_run.run_id]
+    assert persisted_run.run_id is not None
+    assert persisted_run.messages is not None
+    assert "Describe this image" in cast("str", persisted_run.messages[0].content)
+    assert "Available attachment IDs: att_1" in cast("str", persisted_run.messages[0].content)
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_streaming_emits_session_started_after_persisted_delivery_error(
+    tmp_path: Path,
+) -> None:
+    """session:started should still fire for team streams that fail after persisting the session."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = MagicMock(spec=AgentBot)
+    bot.logger = MagicMock()
+    bot.stop_manager = MagicMock()
+    bot.stop_manager.remove_stop_button = AsyncMock()
+    bot.client = AsyncMock()
+    bot.agent_name = "ultimate"
+    bot.storage_path = tmp_path
+    bot.config = config
+    bot.runtime_paths = runtime_paths
+    bot._knowledge_access_support = _knowledge_access_support()
+
+    storage = _SessionStorage()
+    sequence: list[str] = []
+
+    @hook(EVENT_SESSION_STARTED)
+    async def started(ctx: SessionHookContext) -> None:
+        sequence.append(f"started:{ctx.scope.key}:{ctx.session_id}:{ctx.thread_id}")
+
+    registry = HookRegistry.from_plugins([_plugin("session-hooks", [started])])
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=registry,
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        async def consume_delivery_and_fail(request: object) -> StreamTransportOutcome:
+            chunks = [chunk async for chunk in request.response_stream]
+            accumulated = "".join(chunks)
+            sequence.append(f"deliver:{accumulated}")
+            raise StreamingDeliveryError(
+                RuntimeError("boom"),
+                event_id="$team-terminal",
+                accumulated_text=accumulated,
+                tool_trace=[],
+                transport_outcome=_stream_outcome(
+                    "$team-terminal",
+                    accumulated,
+                    terminal_status="error",
+                    failure_reason="boom",
+                ),
+            )
+
+        coordinator.deps.delivery_gateway.deliver_stream.side_effect = consume_delivery_and_fail
+
+        def fake_team_response_stream(*_args: object, **kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                session_id = kwargs["ctx"].session_id
+                assert isinstance(session_id, str)
+                storage.session = TeamSession(
+                    session_id=session_id,
+                    team_id="ultimate",
+                    created_at=1,
+                    updated_at=1,
+                )
+                sequence.append("stream")
+                yield "Team hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+        request = replace(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            existing_event_id="$placeholder",
+            existing_event_is_placeholder=True,
+        )
+
+        resolution = await coordinator.generate_team_response_helper(
+            request,
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$team-terminal"
+    assert sequence == [
+        "stream",
+        "deliver:Team hello",
+        "started:team:ultimate:!test:localhost:$thread-root:$thread-root",
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_streaming_delivery_carries_live_metadata_collector(
+    tmp_path: Path,
+) -> None:
+    """The team streaming delivery request carries the live metadata collector.
+
+    The turn driver fills the collector at terminal settle, before the
+    stream's final edit snapshots extra_content; without the live dict the
+    ai_run payload never reaches Matrix in streaming mode (the finalize
+    happy path sends no extra edit).
+    """
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = MagicMock(spec=AgentBot)
+    bot.logger = MagicMock()
+    bot.stop_manager = MagicMock()
+    bot.stop_manager.remove_stop_button = AsyncMock()
+    bot.client = AsyncMock()
+    bot.agent_name = "ultimate"
+    bot.storage_path = tmp_path
+    bot.config = config
+    bot.runtime_paths = runtime_paths
+    bot._knowledge_access_support = _knowledge_access_support()
+
+    captured_extra_content: list[object] = []
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        async def consume_delivery(request: object) -> StreamTransportOutcome:
+            async for _chunk in request.response_stream:
+                pass
+            # Snapshot after stream exhaustion, like the real final edit.
+            captured_extra_content.append(request.extra_content)
+            return _stream_outcome("$team-final", "Team hello")
+
+        coordinator.deps.delivery_gateway.deliver_stream.side_effect = consume_delivery
+
+        def fake_team_response_stream(*_args: object, **kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                collector = kwargs["run_metadata_collector"]
+                assert isinstance(collector, dict)
+                collector["io.mindroom.ai_run"] = {"usage": {"output_tokens": 5}}
+                yield "Team hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert captured_extra_content == [{"io.mindroom.ai_run": {"usage": {"output_tokens": 5}}}]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_persists_interrupted_history_when_stream_delivery_fails(
+    tmp_path: Path,
+) -> None:
+    """Team stream delivery errors should persist canonical interrupted replay from the partial text."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=HookRegistry.empty(),
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        async def consume_delivery_and_fail(request: object) -> StreamTransportOutcome:
+            accumulated = ""
+            async for chunk in request.response_stream:
+                accumulated += str(chunk)
+            raise StreamingDeliveryError(
+                RuntimeError("boom"),
+                event_id="$team-terminal",
+                accumulated_text="Team hello\n\n**[Response interrupted by an error: boom]**",
+                tool_trace=[],
+                transport_outcome=_stream_outcome(
+                    "$team-terminal",
+                    "Team hello\n\n**[Response interrupted by an error: boom]**",
+                    terminal_status="error",
+                    failure_reason="boom",
+                ),
+            )
+
+        coordinator.deps.delivery_gateway.deliver_stream.side_effect = consume_delivery_and_fail
+
+        def fake_team_response_stream(*_args: object, **_kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                yield "Team hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$team-terminal"
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session is not None
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert persisted_run.messages is not None
+    assert [(message.role, message.content) for message in persisted_run.messages] == [
+        ("user", "Hello"),
+        ("assistant", "Team hello\n\n(turn interrupted by the user before completion)"),
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_stream_delivery_failure_with_visible_tools_replays_tool_trace_once(
+    tmp_path: Path,
+) -> None:
+    """Team stream delivery failures should not persist visible tool markers alongside replay traces."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch.object(ResponseRunner, "_show_tool_calls", return_value=True),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=HookRegistry.empty(),
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        async def consume_delivery_and_fail(_request: object) -> StreamTransportOutcome:
+            raise StreamingDeliveryError(
+                RuntimeError("boom"),
+                event_id="$team-terminal",
+                accumulated_text=(
+                    "🤝 **Team Response** (General):\n\nTeam hello\n\n"
+                    "🔧 `run_shell_command` [1]\n\n"
+                    "**[Response interrupted by an error: boom]**"
+                ),
+                tool_trace=[
+                    ToolTraceEntry(
+                        type="tool_call_completed",
+                        tool_name="run_shell_command",
+                        args_preview="cmd=pwd",
+                        result_preview="/app",
+                    ),
+                ],
+                transport_outcome=_stream_outcome(
+                    "$team-terminal",
+                    (
+                        "🤝 **Team Response** (General):\n\nTeam hello\n\n"
+                        "🔧 `run_shell_command` [1]\n\n"
+                        "**[Response interrupted by an error: boom]**"
+                    ),
+                    terminal_status="error",
+                    failure_reason="boom",
+                ),
+            )
+
+        coordinator.deps.delivery_gateway.deliver_stream.side_effect = consume_delivery_and_fail
+
+        def fake_team_response_stream(*_args: object, **_kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                yield "🤝 **Team Response** (General):\n\nTeam hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$team-terminal"
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session is not None
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assistant_text = cast("str", persisted_run.messages[1].content)
+    assert "🔧 `run_shell_command` [1]" not in assistant_text
+    assert assistant_text.count("run_shell_command") == 1
+    assert assistant_text == (
+        "🤝 **Team Response** (General):\n\nTeam hello\n\n"
+        "(turn interrupted by the user before completion; "
+        "1 tool call(s) had completed: run_shell_command)"
+    )
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_persists_minimal_interrupted_history_after_task_cancel(
+    tmp_path: Path,
+) -> None:
+    """Lifecycle-owned team cancellation should persist one minimal interrupted turn."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = MagicMock(spec=AgentBot)
+    bot.logger = MagicMock()
+    bot.stop_manager = MagicMock()
+    bot.stop_manager.remove_stop_button = AsyncMock()
+    bot.client = AsyncMock()
+    bot.agent_name = "ultimate"
+    bot.storage_path = tmp_path
+    bot.config = config
+    bot.runtime_paths = runtime_paths
+    bot._knowledge_access_support = _knowledge_access_support()
+
+    storage = _SessionStorage()
+    started = asyncio.Event()
+
+    async def fake_run_cancellable_response(**kwargs: object) -> str:
+        response_function = cast("Callable[[str | None], Awaitable[object]]", kwargs["response_function"])
+        task = asyncio.create_task(response_function("$thinking"))
+        await started.wait()
+        task.cancel()
+        with suppress(asyncio.CancelledError):
+            await task
+        return "$thinking"
+
+    with (
+        patch.object(
+            ResponseRunner,
+            "run_cancellable_response",
+            new=AsyncMock(side_effect=fake_run_cancellable_response),
+        ),
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response", new_callable=AsyncMock) as mock_team_response,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(coordinator.deps.delivery_gateway, "edit_text", AsyncMock())
+
+        async def fake_team_response(*_args: object, **_kwargs: object) -> str:
+            started.set()
+            await asyncio.sleep(60)
+            return "unreachable"
+
+        mock_team_response.side_effect = fake_team_response
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$thinking"
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session is not None
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert persisted_run.messages is not None
+    assert persisted_run.messages[0].role == "user"
+    assert "Hello" in cast("str", persisted_run.messages[0].content)
+    assert [(message.role, message.content) for message in persisted_run.messages[-1:]] == [
+        ("assistant", "(turn interrupted by the user before completion)"),
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_persists_interrupted_history_when_final_delivery_is_cancelled(
+    tmp_path: Path,
+) -> None:
+    """Delivery-stage cancellation after a completed team run should still persist replay."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response", new_callable=AsyncMock) as mock_team_response,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "deliver_final",
+            AsyncMock(side_effect=asyncio.CancelledError("delivery cancel")),
+        )
+
+        async def fake_team_response(*_args: object, **kwargs: object) -> str:
+            turn_recorder = cast("TurnRecorder", kwargs["turn_recorder"])
+            turn_recorder.set_run_id("team-run-delivery-cancel")
+            turn_recorder.record_completed(
+                run_metadata={},
+                assistant_text="🤝 Team Response:\n\nTeam hello",
+                completed_tools=[],
+            )
+            return "🤝 Team Response:\n\nTeam hello"
+
+        mock_team_response.side_effect = fake_team_response
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution is None
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session is not None
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert persisted_run.run_id == "team-run-delivery-cancel"
+    assert persisted_run.messages is not None
+    assert [(message.role, message.content) for message in persisted_run.messages] == [
+        ("user", "Hello"),
+        ("assistant", "🤝 Team Response:\n\nTeam hello\n\n(turn interrupted by the user before completion)"),
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_preserves_visible_stream_when_finalize_returns_cancelled(
+    tmp_path: Path,
+) -> None:
+    """Delivery-stage cancellation after team streaming completes should still persist replay."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "finalize_streamed_response",
+            AsyncMock(
+                return_value=FinalDeliveryOutcome(
+                    terminal_status="cancelled",
+                    event_id="$team-msg",
+                    is_visible_response=True,
+                    final_visible_body="Team hello",
+                    delivery_kind="sent",
+                    failure_reason="delivery_cancelled",
+                ),
+            ),
+        )
+
+        async def consume_stream(request: object) -> StreamTransportOutcome:
+            accumulated = ""
+            async for chunk in request.response_stream:
+                accumulated += str(chunk)
+            return _stream_outcome("$team-msg", accumulated)
+
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "deliver_stream",
+            AsyncMock(side_effect=consume_stream),
+        )
+
+        def fake_team_response_stream(*_args: object, **_kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                yield "Team hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$team-msg"
+    assert storage.session is None
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_preserves_structured_stream_cancel_delivery_state(
+    tmp_path: Path,
+) -> None:
+    """Structured team stream cancellation must flow through the dedicated cancelled-delivery path."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response_stream"),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "deliver_stream",
+            AsyncMock(
+                side_effect=StreamingDeliveryError(
+                    asyncio.CancelledError("team stream cancelled"),
+                    event_id="$team-msg",
+                    accumulated_text="Team hello",
+                    tool_trace=[],
+                    transport_outcome=_stream_outcome(
+                        "$team-msg",
+                        "Team hello",
+                        terminal_status="cancelled",
+                        failure_reason="cancelled_by_user",
+                    ),
+                ),
+            ),
+        )
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$team-msg"
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session is not None
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert persisted_run.messages is not None
+    assert [(message.role, message.content) for message in persisted_run.messages] == [
+        ("user", "Hello"),
+        ("assistant", "Team hello\n\n(turn interrupted by the user before completion)"),
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_preserves_visible_stream_on_late_finalize_error(
+    tmp_path: Path,
+) -> None:
+    """Late streamed team finalization errors should preserve the visible stream as an error outcome."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "finalize_streamed_response",
+            AsyncMock(
+                return_value=FinalDeliveryOutcome(
+                    terminal_status="error",
+                    event_id="$team-msg",
+                    is_visible_response=True,
+                    final_visible_body="Team hello",
+                    delivery_kind="sent",
+                    failure_reason="delivery crash",
+                ),
+            ),
+        )
+
+        async def consume_stream(request: object) -> StreamTransportOutcome:
+            accumulated = ""
+            async for chunk in request.response_stream:
+                accumulated += str(chunk)
+            return _stream_outcome("$team-msg", accumulated)
+
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "deliver_stream",
+            AsyncMock(side_effect=consume_stream),
+        )
+
+        def fake_team_response_stream(*_args: object, **_kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                yield "Team hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$team-msg"
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_settles_late_failure_without_finalize(
+    tmp_path: Path,
+) -> None:
+    """Raw late team failures settle a bare terminal error without finalize.
+
+    Mirrors the agent arm: the tracked event must not be routed through the
+    placeholder-only cleanup, because an adopted thinking-message stream can
+    already hold the full streamed reply under the same event id.
+    """
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response_stream"),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "deliver_stream",
+            AsyncMock(side_effect=RuntimeError("stream boom")),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "finalize_streamed_response",
+            AsyncMock(
+                return_value=FinalDeliveryOutcome(
+                    terminal_status="error",
+                    event_id=None,
+                    failure_reason="stream boom",
+                ),
+            ),
+        )
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution is None
+    coordinator.deps.delivery_gateway.finalize_streamed_response.assert_not_awaited()
+    coordinator.deps.delivery_gateway.deps.response_hooks.emit_cancelled_response.assert_awaited_once()
+
+
+def test_record_stream_delivery_error_preserves_hidden_tool_state_when_visible_trace_is_empty(
+    tmp_path: Path,
+) -> None:
+    """Delivery failures must keep hidden tool progress already recorded by the stream generator."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths)
+    coordinator = _build_response_runner(
+        bot,
+        config=config,
+        runtime_paths=runtime_paths,
+        storage_path=tmp_path,
+        requester_id="@alice:localhost",
+    )
+    recorder = TurnRecorder(user_message="Hello")
+    recorder.set_run_metadata({"matrix_seen_event_ids": ["$user_msg"]})
+    recorder.set_assistant_text("Partial answer")
+    recorder.set_completed_tools(
+        [
+            ToolTraceEntry(
+                type="tool_call_completed",
+                tool_name="run_shell_command",
+                args_preview="cmd=pwd",
+                result_preview="/app",
+            ),
+        ],
+    )
+    recorder.set_interrupted_tools(
+        [
+            ToolTraceEntry(
+                type="tool_call_started",
+                tool_name="save_file",
+                args_preview="file_name=main.py",
+            ),
+        ],
+    )
+
+    assert coordinator._record_stream_delivery_error(
+        recorder=recorder,
+        accumulated_text="Partial answer\n\n**[Response interrupted by an error: boom]**",
+        tool_trace=[],
+    )
+
+    snapshot = recorder.interrupted_snapshot()
+    assert snapshot.partial_text == "Partial answer"
+    assert [tool.tool_name for tool in snapshot.completed_tools] == ["run_shell_command"]
+    assert [tool.tool_name for tool in snapshot.interrupted_tools] == ["save_file"]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_persists_original_user_message_for_cancelled_team_run(
+    tmp_path: Path,
+) -> None:
+    """Cancelled team replay should store the raw user turn, not the shaped model prompt."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+    storage = _SessionStorage()
+    model_prompts: list[str] = []
+
+    async def fake_run_cancellable_response(**kwargs: object) -> str:
+        response_function = cast("Callable[[str | None], Awaitable[object]]", kwargs["response_function"])
+        on_cancelled = cast("Callable[[str], None] | None", kwargs.get("on_cancelled"))
+        with suppress(asyncio.CancelledError):
+            await response_function("$thinking")
+        if on_cancelled is not None:
+            on_cancelled("cancelled_by_user")
+        return "$thinking"
+
+    with (
+        patch.object(
+            ResponseRunner,
+            "run_cancellable_response",
+            new=AsyncMock(side_effect=fake_run_cancellable_response),
+        ),
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.teams.Team.arun", new_callable=AsyncMock) as mock_team_arun,
+        patch(
+            "mindroom.teams.open_bound_scope_session_context",
+            side_effect=lambda **_kwargs: _open_team_scope_context(storage),
+        ),
+    ):
+        orchestrator = _team_orchestrator(config, runtime_paths)
+        orchestrator.agent_bots = {"general": SimpleNamespace(running=True)}
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=orchestrator,
+        )
+
+        async def fake_team_arun(prompt: str, **kwargs: object) -> TeamRunOutput:
+            model_prompts.append(prompt)
+            return TeamRunOutput(
+                run_id=cast("str | None", kwargs.get("run_id")),
+                team_id="ultimate",
+                session_id=cast("str | None", kwargs.get("session_id")),
+                content="Run cancelled",
+                messages=[Message(role="assistant", content="Run cancelled")],
+                status=RunStatus.cancelled,
+            )
+
+        mock_team_arun.side_effect = fake_team_arun
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$thinking"
+    assert model_prompts
+    assert model_prompts[0] != "Hello"
+    assert 'Current message:\n<msg from="@alice:localhost">' in model_prompts[0]
+    assert "Hello" in model_prompts[0]
+    persisted_session = cast("TeamSession", storage.session)
+    assert persisted_session is not None
+    assert persisted_session.runs is not None
+    persisted_run = cast("TeamRunOutput", persisted_session.runs[0])
+    assert persisted_run.messages is not None
+    assert [(message.role, message.content) for message in persisted_run.messages] == [
+        ("user", "Hello"),
+        ("assistant", "(turn interrupted by the user before completion)"),
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_emits_session_started_after_persisted_cancellation(
+    tmp_path: Path,
+) -> None:
+    """session:started should still fire when a cancelled team run has already persisted the session."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = MagicMock(spec=AgentBot)
+    bot.logger = MagicMock()
+    bot.stop_manager = MagicMock()
+    bot.stop_manager.remove_stop_button = AsyncMock()
+    bot.client = AsyncMock()
+    bot.agent_name = "ultimate"
+    bot.storage_path = tmp_path
+    bot.config = config
+    bot.runtime_paths = runtime_paths
+    bot._knowledge_access_support = _knowledge_access_support()
+
+    storage = _SessionStorage()
+    sequence: list[str] = []
+
+    @hook(EVENT_SESSION_STARTED)
+    async def started(ctx: SessionHookContext) -> None:
+        sequence.append(f"started:{ctx.scope.key}:{ctx.session_id}:{ctx.thread_id}")
+
+    registry = HookRegistry.from_plugins([_plugin("session-hooks", [started])])
+
+    async def fake_run_cancellable_response(**kwargs: object) -> str:
+        response_function = cast("Callable[[str | None], Awaitable[object]]", kwargs["response_function"])
+        on_cancelled = cast("Callable[[str], None] | None", kwargs.get("on_cancelled"))
+        with suppress(asyncio.CancelledError):
+            await response_function("$thinking")
+        if on_cancelled is not None:
+            on_cancelled("cancelled_by_user")
+        return "$thinking"
+
+    with (
+        patch.object(
+            ResponseRunner,
+            "run_cancellable_response",
+            new=AsyncMock(side_effect=fake_run_cancellable_response),
+        ),
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response", new_callable=AsyncMock) as mock_team_response,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=registry,
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(coordinator.deps.delivery_gateway, "edit_text", AsyncMock())
+
+        async def fake_team_response(*_args: object, **kwargs: object) -> str:
+            cancel_message = "cancel"
+            session_id = kwargs["ctx"].session_id
+            assert isinstance(session_id, str)
+            storage.session = TeamSession(
+                session_id=session_id,
+                team_id="ultimate",
+                created_at=1,
+                updated_at=1,
+            )
+            sequence.append("team")
+            raise asyncio.CancelledError(cancel_message)
+
+        mock_team_response.side_effect = fake_team_response
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution == "$thinking"
+    assert sequence == [
+        "team",
+        "started:team:ultimate:!test:localhost:$thread-root:$thread-root",
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_streaming_emits_session_started_after_persisted_cancellation(
+    tmp_path: Path,
+) -> None:
+    """session:started should still fire when a cancelled team stream has already persisted the session."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = MagicMock(spec=AgentBot)
+    bot.logger = MagicMock()
+    bot.stop_manager = MagicMock()
+    bot.stop_manager.remove_stop_button = AsyncMock()
+    bot.client = AsyncMock()
+    bot.agent_name = "ultimate"
+    bot.storage_path = tmp_path
+    bot.config = config
+    bot.runtime_paths = runtime_paths
+    bot._knowledge_access_support = _knowledge_access_support()
+
+    storage = _SessionStorage()
+    sequence: list[str] = []
+
+    @hook(EVENT_SESSION_STARTED)
+    async def started(ctx: SessionHookContext) -> None:
+        sequence.append(f"started:{ctx.scope.key}:{ctx.session_id}:{ctx.thread_id}")
+
+    registry = HookRegistry.from_plugins([_plugin("session-hooks", [started])])
+
+    async def fake_run_cancellable_response(**kwargs: object) -> str:
+        response_function = cast("Callable[[str | None], Awaitable[object]]", kwargs["response_function"])
+        on_cancelled = cast("Callable[[str], None] | None", kwargs.get("on_cancelled"))
+        with suppress(asyncio.CancelledError):
+            await response_function("$thinking")
+        if on_cancelled is not None:
+            on_cancelled("cancelled_by_user")
+        return "$thinking"
+
+    with (
+        patch.object(
+            ResponseRunner,
+            "run_cancellable_response",
+            new=AsyncMock(side_effect=fake_run_cancellable_response),
+        ),
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=registry,
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        async def consume_delivery_and_cancel(request: object) -> StreamTransportOutcome:
+            accumulated = ""
+            async for chunk in request.response_stream:
+                accumulated += str(chunk)
+                sequence.append(f"deliver:{accumulated}")
+            return _stream_outcome("$team-msg", accumulated)
+
+        coordinator.deps.delivery_gateway.deliver_stream.side_effect = consume_delivery_and_cancel
+
+        def fake_team_response_stream(*_args: object, **kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                cancel_message = "cancel"
+                session_id = kwargs["ctx"].session_id
+                assert isinstance(session_id, str)
+                storage.session = TeamSession(
+                    session_id=session_id,
+                    team_id="ultimate",
+                    created_at=1,
+                    updated_at=1,
+                )
+                sequence.append("stream")
+                yield "Team hello"
+                raise asyncio.CancelledError(cancel_message)
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution is None
+    assert sequence == [
+        "stream",
+        "deliver:Team hello",
+        "started:team:ultimate:!test:localhost:$thread-root:$thread-root",
+    ]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_uses_persisted_team_scope_for_session_started_hooks(
+    tmp_path: Path,
+) -> None:
+    """Ad hoc team session hooks should scope to the persisted team scope, not the router bot."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name=ROUTER_AGENT_NAME)
+
+    storage = _SessionStorage()
+    sequence: list[str] = []
+
+    @hook(EVENT_SESSION_STARTED, agents=["team_general"], rooms=["!test:localhost"])
+    async def started(ctx: SessionHookContext) -> None:
+        sequence.append(f"started:{ctx.scope.key}:{ctx.agent_name}")
+
+    registry = HookRegistry.from_plugins([_plugin("session-hooks", [started])])
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_runner.team_response", new_callable=AsyncMock) as mock_team_response,
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=registry,
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        async def fake_team_response(*_args: object, **kwargs: object) -> str:
+            session_id = kwargs["ctx"].session_id
+            assert isinstance(session_id, str)
+            storage.session = TeamSession(
+                session_id=session_id,
+                team_id="team_general",
+                created_at=1,
+                updated_at=1,
+            )
+            return "Team hello"
+
+        mock_team_response.side_effect = fake_team_response
+
+        await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert sequence == ["started:team:team_general:team_general"]
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_merges_raw_prompt_into_model_prompt(
+    tmp_path: Path,
+) -> None:
+    """Ad hoc team responses should keep the user request when model_prompt only adds metadata."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name=ROUTER_AGENT_NAME)
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=False)),
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+        patch("mindroom.response_runner.team_response", new_callable=AsyncMock) as mock_team_response,
+    ):
+        mock_team_response.return_value = "Team hello"
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(
+                prompt="What is in the image?",
+                model_prompt="Available attachment IDs: att_img. Use tool calls to inspect or process them.",
+                user_id="@alice:localhost",
+                thread_id="$thread-root",
+            ),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert _handled_response_event_id(resolution) == "$thinking"
+    assert mock_team_response.await_args is not None
+    message = mock_team_response.await_args.kwargs["message"]
+    assert "What is in the image?" in message
+    assert "Available attachment IDs: att_img. Use tool calls to inspect or process them." in message
+
+
+@pytest.mark.asyncio
+async def test_generate_team_response_helper_uses_delivery_result_failure_reason_for_cancelled_stream(
+    tmp_path: Path,
+) -> None:
+    """Typed gateway outcomes should preserve their canonical failure_reason."""
+    runtime_paths = _runtime_paths(tmp_path)
+    config = bind_runtime_paths(_config_with_team(), runtime_paths)
+    bot = _make_bot(tmp_path, config=config, runtime_paths=runtime_paths, agent_name="ultimate")
+
+    storage = _SessionStorage()
+
+    with (
+        patch("mindroom.response_runner.should_use_streaming", new=AsyncMock(return_value=True)),
+        patch("mindroom.response_runner.team_response_stream") as mock_team_stream,
+        patch("mindroom.response_lifecycle.apply_post_response_effects", new=AsyncMock(return_value=None)),
+    ):
+        coordinator = _build_response_runner(
+            bot,
+            config=config,
+            runtime_paths=runtime_paths,
+            storage_path=tmp_path,
+            requester_id="@alice:localhost",
+            hook_registry=HookRegistry.empty(),
+            team_history_storage=storage,
+            message_target=MessageTarget.resolve("!test:localhost", "$thread-root", "$user_msg"),
+            orchestrator=_team_orchestrator(config, runtime_paths),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "deliver_stream",
+            AsyncMock(
+                return_value=StreamTransportOutcome(
+                    last_physical_stream_event_id="$team-msg",
+                    terminal_status="completed",
+                    rendered_body="Team hello",
+                    visible_body_state="visible_body",
+                ),
+            ),
+        )
+        _set_gateway_method(
+            coordinator.deps.delivery_gateway,
+            "finalize_streamed_response",
+            AsyncMock(
+                return_value=FinalDeliveryOutcome(
+                    terminal_status="error",
+                    event_id=None,
+                    failure_reason="stream failure",
+                ),
+            ),
+        )
+
+        def fake_team_response_stream(*_args: object, **kwargs: object) -> AsyncIterator[str]:
+            async def fake_stream() -> AsyncIterator[str]:
+                session_id = kwargs["session_id"]
+                assert isinstance(session_id, str)
+                storage.session = TeamSession(
+                    session_id=session_id,
+                    team_id="ultimate",
+                    created_at=1,
+                    updated_at=1,
+                )
+                yield "Team hello"
+
+            return fake_stream()
+
+        mock_team_stream.side_effect = fake_team_response_stream
+
+        resolution = await coordinator.generate_team_response_helper(
+            _response_request(prompt="Hello", user_id="@alice:localhost", thread_id="$thread-root"),
+            team_agents=[fixture_entity_matrix_id("general", "localhost", runtime_paths)],
+            team_mode="coordinate",
+        )
+
+    assert resolution is None
